Removed --enable-audiotoolbox and --enable-vda
This will fix errors when running the configure script. These
components are only for Apple systems and does not need to be
enabled in Linux anyway.

Also removed the initial comment section about the not
available options.

The errors in configure script were caused by the following
upstream git commit:
http://git.videolan.org/?p=ffmpeg.git;a=commit;h=8aad209c13c2a66a9256e37d7f44b5f9f19b13b2