flite dependency changes
Changed flite-fpic to flite1. flite-fpic uses static libraries and
flite1 uses shared ones, so flite1 is preferred.

It seems that flite is needed only at building time since it's not
linked in any binaries and namcap does not complain about it.
flite functionality in ffmpeg and ffplay works as expected when
deleting flite1 after installing ffmpeg-full-git.

Also, the flite detection code on prepare() seems to be not needed
since everything works without it.

Note: flite (2.x) that is present at [community] repository is not
supported by ffmpeg.