Mature the package.
* Make the daemon work better
This installs a working systemd config file so users can just call
   systemd start bitcoind
and a simple way to communicate with that deamon using the cli util.

* Compile it;
Instead of downloading the pre-compiled binaries which are statically
linked, compile the software from the source release to avoid
the binary being larger than needed due to those extra libraries
it linked in.