Create user systemd units
Those are a better solution for the audio issues, since PulseAudio
itself does not recoomend to get it running as a system instance and
Gnome also is not able to manage the audio preferences with such
configuration.

The units provide a standard interface to starting mycroft services
allowing even multiple user installations each one paired independently.