Updated version (r4009.69e0ebe6 -> r4013.6cc2f04b).
Changelog
=========
These two accelerators apparently don't want to be translated. (6cc2f04b)
Better documentation of the speedUp. (0af90985)
Better handling of the last newline of a text cell. (185dbe8e)
This should get rid of most travis build errors for mac: The travis build cloned only the last 3 commits to the repository, switched to the commit that was about to be tested - and sometimes testing starts up to an hour after the commit meaning that the about-to-be-tested version no more is part of the 3 commits travis' working copy can be reverted to. (1b392934)