Merge patch from @KingCrunch
* Add missing LICENSE-file
* Add .gitignore
* Fix supported arm-architectures
placed arm woth armv6h and amv7h. arm refers to ARM5, which doesn't support hardfloat,
thus the armhf-package provided by upstream will probably not work.
* Add provided LICENSE to PKGBUILD
* Add conflict with `telegraf`
because it makes use of the same files and folders
* Create /var/log/telegraf
because it is "somehow suggested" by the upstream package
* Add logrotate-configuration
makes sense to use when we expect log files
* Use service-file provided by upstream-package
* Copy provided LICENSE to correct location
* Don't use removed service-file
* Credit @KingCrunch for patch
* Increase pkgrel to 2
* Generate .SRCINFO