updpkg: 0.84 + remove bundled libs
The bundled libs are broken in 0.84, some Qt 5 libraries are 5.6.1 and
others 5.6.2, the former depending on icu 54 and the latter on icu 56,
which is not even provided.

Since it actually works with the system libs, let's remove all that
mess. It may break at some point, but at least for now it's the easiest
option (and it's also much better to use the system libs).

Also added some cleanup (we use our own desktop file).