Version source archives
Add the version number to the downloaded source archives. This talks
makepkg into redownloading the sources when there is a new release,
instead of keeping the old archive and failing the checksum. This also
has the nice side-effect of keeping previous source versions, should one
want to downgrade the package.

Since upstream does not always bump the version number when releasing
new sources, I had to use a trick to make it work. "-rel$srcrel" is
appended to the filename, and if new sources of the same version are
released, srcrel is incremented (and then reset to 1 when a new version
is released). Not beautiful, but it should do the job.