Removed oss from optdepends
oss was wrongly placed only in optdepends. To enable oss support in
wine it needs to be installed at build time and thus would require
it to be placed at depends. The oss AUR package removes by default
the alsa kernel modules, which is not a good thing. Since most users
rely on alsa, it is really unconfortable to maintain both alsa and
oss installed on the same system.