pkgver() must not produce any superfluous output.
Usually, makepkg just takes the stuff that pkgver() prints to stdout, so
we can output any log messages to stderr. However, when makepkg is in
logging mode ("makepkg -L ..."), then it merges stdout and stderr
together and uses the result as the pkgver (I'm sure that's a bug!),
which breaks the build horribly due to us logging things. So do not log
anything at all now.