tibia: Update to build 11.10.4445 and some cleanup
Give the agreement file a name, which can be associated with a package. This is for people using a common SRCDEST for alle Source packages.

Problem which needs to be resolved is for people which use BUILDDIR. gendesk seems to hardcode the path for the PKGBUILD, which may not be the upper directory