updpkg: 0.84-3
So apparently setting QT_QPA_PLATFORMTHEME was not enough to fix the
font issue. I found out that using QT_PLUGIN_PATH=/usr/lib/qt/plugins
didn't work for me because of an obscure issue in the OpenGL xcb
integration. So my idea was to use the system Qt plugins, but not the
whole directory as some stuff doesn't work. Symlinking just platforms
and platformthemes seems to do the trick: no segfault and no font
problem.

Do keep in mind that this package is experimental and things may break
as I am experimenting. But of course, when this happens, let me know and
I'll try my best to make it work for everyone.