depends: added ghostscript and temporarily removed libemf
ghostscript is a dependency and was missing. Although namcap says
that it's not needed, configure script checks for gs and gives an
error if it's not found.

libemf is currently outdated in AUR (1.0.8) and this version has a
bug that places the emf.h header file in the wrong directory,
making it unuseable. Using CPPFLAGS in pstoedit configure script
doesn't work for detecting the emf.h in 1.0.8. Version 1.0.9 fixes
this bug. When the AUR package get updated we can bring back libemf
support.

Also, removed the libemf source download since it seems completely
unecessary.