depends: changed back to zimg (from zimg-git, no more compile error)
In a previous AUR git commit from 2016-11-26 it was needed to change
zimg dependency to zimg-git to fix a compile error. It seems that
ffmpeg git master was (is?) using code from zimg git master. Now that
a new zimg version (v2.4) was released a while ago there is no more
compile error.

If there will be another similiar issue in the future of course we
can consider to change this again.

Reference:
https://aur.archlinux.org/cgit/aur.git/commit/?h=ffmpeg-full-git&id=8fed268043cb2c3fcbcab550c429caf4e7f0e300