Upgraded to version 0.29.0; updated sha256sums accordingly; removed ${pkg_base}.install; disabled compile flag removal;
The old versions of PKGBUILD contained the following lines in prepare():

    # remove specific build flags so that the options from makepkg.conf are honored
    sed -i 's/-O2//g' configure
    sed -i 's/-D_FORTIFY_SOURCE=2//g' configure
    sed -i 's/-DNDEBUG//g' configure
    sed -i 's/-I\/usr\/local\/include//g' configure

IMHO sed-patching the configure script is a bad idea, especially if not needed.

  -O2  is superseeded by whatever is specified in makepkg.conf anyway
  -D_FORTIFY_SOURCE  is not harmfull, especially if tested by the upstream developers
  -DNDEBUG  is a reasonalbe choice for most users. Especially as many arch users
            do not specify it explicitly in their makepkg.conf
  -I/usr/local/include typically has no effect anyway, since it is empty
                       by default on arch linux.

So I consider it less harmfull to stay with the default ./configure provided by
the developers than patching it on the fly. If a user really wants to change the
build behaviour, she can do this trough makepkg more elegantly.