Republish the package with developer keyword
This avoids a naming conflict between the dependent packages:

- nvidia-utils-vulkan-developer-beta
- nvidia-libgl-vulkan-developer-beta
- opencl-nvidia-vulkan-developer-beta
- lib32-nvidia-utils-vulkan-developer-beta
- lib32-nvidia-libgl-vulkan-developer-beta
- lib32-opencl-nvidia-vulkan-developer-beta
- nvidia-vulkan-developer-beta

The original packages were named without the keyword and pushed, but
that was before I was stopped while trying to push the
"nvidia-vulkan-beta" into an already existing repository without a
corresponding package (404). So I added the developer keyword to all 7
packages. Sorry for the mess!