Updated version (r4087.c1a30c58 -> r4089.d22856da).
Changelog
=========
Allow brackets to be hidden as long as they don't provide information that is currently used. Looks more modern, but still doesn't seem to improve the usability => This feature is inactive by default. This is the second half of the solution of #857 (d22856da)
A case in which a single letter in text cells might disappear. (f49533e4)