Move sdl_net to depends, use sha512sums instead of sha256sums.
sdl_net is required to build, so this package failed to build on
a clean system (eg, in ccm).  Could have moved it into makedepends,
but it was easy enough this way.

sha512 is faster to compute than sha256.  The difference is
negligible for Chocolate Doom, but I'll standardize all my packages
on it.