Added support for umem (libumem)
Although configure option --with-umem was specified, there was no
umem package available in repositories or in the AUR that can be
used as a dependency to give compilation support. As a result,
umem was not being detected at configure time and the package was
without umem suppport. Now that libumem-git is available in the
AUR the support can finally be done.