Updated version (r4089.d22856da -> r4092.513539b7).
Changelog
=========
Deactivated autowrap for code. When  - the line was wider than the screen  - and autowrap increased the indentation far enough to make the line *longer* on autowrap something took wrong decisions. (513539b7)
Printing: Use the whole paper width. (65720b40)
A context menu that allows to fold cells providing an alternative workflow to the brackets. (c3aff869)