Changes to the _install_license() logic
So far we've simply installed the LICENSE.TXT file located in the LLVM
root as the sole license file. However, as seen in #10, there are other
license files too. In fact, there are also license files in some of the
subdirectories, which is also indicated in the LICENSE.TXT itself. The
new `_install_licenses()` tries to be a tiny bit smart about the issue
by `find`-ing all appropriate license files.

Please note that the new function does require a path as an argument.