Added patches with various fixes. Removed autotrace fix.
Some fixes are needed. See references bellow.

Also, removed the autotrace fix (that makes autotrace detect new
versions of pstoedit like 3.70) because it's now fixed directly
in package autotrace-nomagick as of its version 0.31.1-3. This
is more logic due to the fact that this bug resides in autotrace
and not in pstoedit.

References
==========

Send errors to stdout
---------------------
http://bugs.debian.org/551125

Fix for pkg-config (obsolete LIBPNG_LDFLAGS)
--------------------------------------------
https://bugs.gentoo.org/show_bug.cgi?id=551208
https://bugzilla.redhat.com/show_bug.cgi?id=1415258

Fix plugin loading
------------------
http://bugs.debian.org/813316