The redistribuable seems to be "custom licensed".
As the redistribuable contains a custom license witch states that
VoxelShop is a Freeware, it should be mark as custom licensed. Even if
the GitHub repository constains an Apache license file.

It is certainly due to the tools used to make the redistribuable.
Confirmation was asked by e-mail to the software owner.