expand release file syntax
Files appeared that use an underscore instead of a hyphen, so the
patterns have to adapt to find the correct version number again.

As the still noted version 1.9.8 can be parsed using the old pattern,
this does not trigger a pkgrel change. But it enables the next pkgver
update to begin with.