do not install modes
In PKGBUILDs, files are normally installed with the install program to
control their permissions. As it cannot recursively process directories,
cp, which can, is instead configured to act similarly.

Resetting modes would rob the executable from its executable bit, so
this single file is instead installed using install which uses the
correct mask of 755 by default.