include dedicated starter script
As the desktop file, this could be generated from a template.

The old script did never work correctly, anyway, and it did for two
reasons: When the CARCH modification was still in the prepare function,
packaging would configure it to search for a executable ending in .i686
while unity uses .x86 instead. Additionally, the data directory is only
found in the current working directory, so a cd is necessary.