Initial Commit
Since the origin gpgme pacakge only provides the bindings
for C++, C and QT, I thought it would be useful to add
the python bindings as well. To keep the update level
as low as possible I decided to use the git repo as
a source, and not the tarball, like the original
gpgme package.