Added check() as a comment block
It takes really a lot of time to run the tests (it likely triples
the building time in a i7 4.0GHz + GTX 750 machine) and it will
prevent the package to be built in case of error. So it is not
enabled by default and is added as a comment block. Users that
need to make the tests should manually enable it. At the time of
this writing I'm getting a 100% success tests.

You can say that there is makepkg '--nocheck' option but the
intention here is to not enable tests by default. Still, many
people can forget about the makepkg '--nocheck' option in case
of failure.