Split the directory structure into two
To use previous versions, one should copy the whole bunch of the
subdirectory under /opt/hifive1-sdk, which is a duplication because
in theory, the toolchain and binary utils doesn't have to be moved.

After this patch, the user now can copy a smaller amount of required
files: the example applications, the platform-specific settings, and the
makefile which plays the main role of the development flow.

Also, the message after installation is more comprehensive now.