First commit after package adoption. Made working again.
Package is now working again.
slimit2 AUR package is now needed as a makedepend.

Important note about libtorrent-rasterbar dependency:
-----------------------------------------------------
deluge-git currently has an upstream requirement of a specific
libtorrent-rasterbar version, which is 1.1.2. It will not work
with libtorrent-rasterbar 1.1.1 or the current
libtorrent-rasterbar git master. But the libtorrent-rasterbar
version offered in the binary repositories is currently outdated,
being at 1.1.1.

Being such, you have 3 options if you want to build and use
deluge-git:

1) use libtorrent-rasterbar-git 1.1.2.r0.ga42d4390c, which is the
default PKGBUILD method. This is the same as libtorrent-rasterbar
1.1.2, but using the git repository. For this you will need to
manually modify libtorrent-rasterbar-git PKGBUILD in order to
obtain this specific version: append '#commit=a42d4390c' at the
source line (currently there is a ready to use source line for
achieving this in libtorrent-rasterbar-git PKGBUILD).

2) use libtorrent-rasterbar 1.1.2. For this you will need to use
the PKGBDUILD from the binary repository and modify it to get
version 1.1.2.

3) wait for the release of libtorrent-rasterbar 1.1.2 in binary
repositories.

This situation will be updated when the binary repositories
upgrade libtorrent-rasterbar to 1.1.2 or when the develop branch
of deluge change/update this requirement.