Replaced libz dependency with zlib
Not sure why it was like that before but I tested using the default zlib
provided by ArchLinux ARM and it works fine. This will probably be very
useful on systems that depends of zlib and couldn't install it because
of conflict dependency issues.

I noticed the problem after a system upgrade because libpng now
requires a more recent version of zlib.