A fully automated build/install with AUR helpers is now supported
In short:
---------
ffmpeg-full-git can now have a fully automated build/install with AUR
helpers! This was not possible previously because some dependencies
could only be installed manually with plain makepkg, but now this is
solved. Tested and currently working with pacaur and yaourt in x86_64.

The long story:
---------------
This package has dependencies that _previously_ did not support AUR
helpers and could only be built manually by running plain makepkg:
  1) blackmagic-decklink-sdk, which needed a manual source file
     download from the upstream website;
  2) flite1, which had problems when building with the popular yaourt
     (although it could be normally built with other AUR helpers like
     pacaur and packer).

As a result, ffmpeg-full-git also was not supporting AUR helpers and
could only be built manually with plain makepkg. This situation was
very sad.

Fortunately, now this is solved:
  1) blackmagic-decklink-sdk as of version 10.8.3-2 does not need
     manual source file download anymore and can be build with AUR
     helpers;
  2) flite1 was "fixed" to allow building with yaourt.

It means that everything that was preventing ffmpeg-full-git from
being used by AUR helpers is solved. Now this package can have a
fully, 100% automated build/install with AUR helpers. Tested and
currently working with pacaur and yaourt in x86_64.

packer still cannot build ffmpeg-full-git because it does not support
dependencies with architecture-specific names. I have made a pull
request in packer GitHub webpage (#156) that fixes this, and it is
waiting for the developer to merge it. One can use the patch from
this pull request in order to make packer to build ffmpeg-full-git.

This AUR git commit is just a marker. Cheers...

References:
-----------
https://aur.archlinux.org/cgit/aur.git/commit/?h=blackmagic-decklink-sdk&id=67b2558b34166f3ef448420ddaa35b74f3007aef
https://aur.archlinux.org/cgit/aur.git/commit/?h=flite1&id=5dd3596cf67ddf105834e8817cd41ac59b6e21ef
https://github.com/keenerd/packer/pull/156