Upgrade to version 1.3.0.
The package also now includes mlmmj's documentation text files in
/usr/share/doc/mlmmj.

A licensing information error was fixed; mlmmj is licensed under the
MIT license, not under GPL. This was incorrectly tagged in the
metadata before. The 'license' field has been fixed accordingly and
the license files are now included in the package as well.