Minimal changes to use Oracle JRE
This fixes the reported bug where launch halts at the "Processing
startup parameters" stage.

It now appears Oracle JRE is required. Some testing was performed with
the java-openjfx package, which allowed launch but no interaction with
the resulting application window was available. It is noted the ib-gw
script to load IB Gateway appears to work without needing the Oracle
JRE, but separating the packages is of limited value given the
relatively limited number of expected deployments and the already high
maintenance cost of this package.