New upstream option: --enable-cuda-sdk
In short:
---------
nvidia-utils is needed as a dependency to enable cuda_sdk features.

The long story:
---------------
ffmpeg configure script defines that cuda_sdk requires libcuda.so
(-lcuda). Two packages provides it: cuda and the nvidia-utils
family. Note that cuda provides only libcuda.so and nvidia-utils
provide the more complete set of libraries with so-version number,
in this case libcuda.so.1.

Binaries from the resulting compilation links to libcuda.so.1
even if we try to use libcuda.so from cuda package. When trying
to build using libcuda.so from cuda package, namcap gives this
warning: "Referenced library 'libcuda.so.1' is an uninstalled
dependency". As fas as I can tell, this warning is harmless, but
it should happen only for libraries provided by the package that
is been compiled (ffmpeg-full-git).

The only package providing the needed file libcuda.so.1 is the
nvidia-utils family. This  library name and so-version is strictly
defined in ffmpeg source code.

Of course this can be revised in future if some other clue is
found.

References:
-----------
http://git.videolan.org/?p=ffmpeg.git;a=commit;h=6a3740572dfeef2628a1c54ba61e7e413bfb55f8
https://github.com/FFmpeg/FFmpeg/blob/edf686f089d68092c3b17a23cc48667665b5a069/configure#L5772
https://github.com/FFmpeg/FFmpeg/blob/edf686f089d68092c3b17a23cc48667665b5a069/compat/cuda/dynlink_loader.h#L54