Fix build if package is installing cuda for the first time
When installing cuda for the first time, the cuda binaries appear
on PATH only after the user do a relogin.

As a consequence, the cuda binaries will not be available if cuda
is being installed for the first time as a dependency of
ffmpeg-full-git. This can be fixed by strictly specifying the full
path of nvcc.

Reference:
----------
https://git.archlinux.org/svntogit/community.git/tree/trunk/cuda.install?h=packages/cuda