license: changed to just nonfree and unredistributable
Although GPL3 is present in the resuling complex license mix, it
does not seem to make sense leaving GPL3 in the license list. The
output of configure script clearly says that the resulting license
is nonfree and unredistributable, and this is what should be
clearly stated to the user.

The ffmpeg documentation says that in such cases, like we are using
here with --enable-nonfree, the resulting binaries are under a
complex and restrictive license mix. But it does not provide clear
licensing terms for such situation. So I just created a LICENSE
file and added generic notes to it which seems to better describe
this nature than the previously used UNREDISTRIBUTABLE.txt file.
Also, the presence of a LICENSE file is something more standard and
is described in the Wiki.

References:
-----------
https://github.com/FFmpeg/FFmpeg/blob/a47273c803edfbc43793349b74429ae29b05c003/configure#L101-L102
https://github.com/FFmpeg/FFmpeg/blob/a47273c803edfbc43793349b74429ae29b05c003/configure#L6554-L6555
https://github.com/FFmpeg/FFmpeg/blob/a47273c803edfbc43793349b74429ae29b05c003/LICENSE.md#incompatible-libraries
https://ffmpeg.org/pipermail/libav-user/2012-November/003179.html