Removed manual GNOME icon cache update as suggested in the comments, however I did not add gtk-update-icon-cache to the dependencies since it is already a dependency of webkit2gtk (as per the wiki guidelines).
Also removed libxtst dependency. As far as I could track it, it was added to eclipse-common in response to bug FS#29445 and later removed without any comments, so I'm removing it from here too.

I'm also not sure why unzip is a dependency. I tried it without unzip on a fresh VM and it seems to work fine, but I decided to leave it since eclipse-common still depends on it.