Recent toolchain changes rebuild
Packages that uses hardening-wrapper will need a rebuild without
it since it is being dropped from official repositories due to
recent toolchain changes (PIE and SSP enabled by default in
gcc). Although hardening-wrapper was not being used in this
package it seems a good practice to rebuild against the recent
toolchain changes.

References:
-----------
https://lists.archlinux.org/pipermail/arch-dev-public/2017-July/028909.html
https://www.archlinux.org/todo/hardening-wrapper-removal/