PIE static libraries rebuild / Recent toolchain changes rebuild
The recent toolchain changes (PIE and SSP enabled by default in
gcc) makes the need to rebuild packages that uses static
libraries as makedepends. A static library that is in makedepends
of ffmpeg-full-git triggered a rebuild (libmfx).

Also, packages that uses hardening-wrapper will need a rebuild
without it since it is being dropped from official repositories
due to the recent toolchain changes. The ffmpeg from oficial
repositories uses hardening-wrapper. Although hardening-wrapper
was not being used in ffmpeg-full-git, it seems a good practice
to rebuild against the recent toolchain changes.

References:
-----------
https://lists.archlinux.org/pipermail/arch-dev-public/2017-July/028909.html
https://www.archlinux.org/todo/pie-rebuild/
https://www.archlinux.org/todo/hardening-wrapper-removal/