Adjust for pacaur
Turns out that Pacaur does more than one pass through the PKGBUILD,
so the timestamp used to avoid pacaur's overzealous build caching
is not matching on the later run!

Thus instead we pre-emptively fetch the txt file that contains the
nightly *build* date, and use that to identify the downloaded file.