Forced building against openssl-1.0
I am not happy with the solution though as I haven't found a clean
way of forcing openssh to build against a specific openssl version.

I tried pkg-config, but either it doesn't work or I am doing it
wrong. Now I am using a hack, where I am using prepare() to symlink
the openssl-1.0 files into the build directory and pass them to the
configure script via the --with-ssl-dir option.

If anyone has a cleaner way of doing this, please feel free to share
it. Any help is greatly appreciated.