Updated to latest version.
Updated to latest upstream version. Small caveat:

The bug described in pr79242 is still not resolved and activity on it seems to
be of low interest. Unfortunately, I do not have the time to look into it that
much. As such, instead of delaying the update until the bug is resolved, I have
opted to instead revert the change that exposed the bug [2]. Yes, this means
that the bug is still there and may be triggered a different way, I have yet to
find a scenario where it happens though.

I will ask then the community: if such case is found, please let me know and/or
add it to [1]. I will then prepare and release a PKGBUILD based on gcc 6.4
instead.

Cheers,
Orlando.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=79242
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71289