Update sha256sum
This hash change is due to the use of gitattributes in the file
ivre/__init__.py. Just after the release, the string '$Format:%h %D$'
is evaluated to '[COMMIT HASH] HEAD -> master, tag: v[VERSION]'. When
a new commit is pushed to master, the same string is evaluated to
'[COMMIT HASH] tag: v[VERSION]'.

The result is exactly the same, since 'tag: v0.9.6' is always present
(see diff below for some context), but this causes the sha256sum to
differ.

--- orig/ivre-0.9.6/ivre/__init__.py    2017-06-26 19:10:14.000000000 +0200
+++ new/ivre-0.9.6/ivre/__init__.py     2017-06-26 19:10:14.000000000 +0200
@@ -72,7 +72,7 @@ def _version():
             return fdesc.read()
     except IOError:
         pass
-    hashval, refnames = u'e5dff0d7 HEAD -> master, tag: v0.9.6'.split(' ', 1)
+    hashval, refnames = u'e5dff0d7 tag: v0.9.6'.split(' ', 1)
     try:
         return next(ref[6:] for ref in refnames.split(u', ') if ref.startswith(u'tag: v'))
     except StopIteration: