Add openrc-script alternative when user has OpenRC installed
This avoid the whole problem to remove the systemd service
installed by default and write by yourself a OpenRC script.

Beyond that, I remove the trailing spaces of the PKGBUILD file,
fix the formatting and add triggers on post_install and pre_remove
about OpenRC.

I tested on my own machine and this seems works fine.