Updated version (r4849.3d5dba6b -> r4852.6536b32c).
Changelog
=========
Support cells ending in a comment that is preceded by the $ or ;. (6536b32c)
SpeedUp: Only redraw the cell contents if the draw region expands enough to the left that it hits it. (aca33541)
SpeedUp: Remove the "force update" flag as soon as it is no more needed. (19c31663)