Update to v1.1.0
Using a different approach to install the package. Instead of download
only the complete font files, downloading all the source code and
preparing it to build the package.

This approach helps the maintenance of the package easier. Now, each
time updating the package, the maintainer only needs to change the
package version number.