Updated version (r4864.6256b147 -> r4874.112c24b1).
Changelog
=========
Populate the table of contents as soon as it is enabled. (112c24b1)
The increased svg resolution allows us to use less safety for parenthesis part heights. (494ff2c5)
Updated the ChangeLog. (e1d8ae20)
Updated the translations. (0bf0ee52)
The "english" translation was still empty and wxmaxima complained about that. Also I am no more sure that it made sense in the first place: On the plus side it allows to correct small errors without marking the translations of the string as "fuzzy". But on the other hand this automatically means that we leave small errors in the code and in the strings for the translations - which doesn't make too much sense. (23b9a264)
Corrected the licenses of 2 Tango icons so debian won't complain about them being set up The Wrong Way. (750eab3d)
This file might prove useful one day. (84ae1522)
Try to make cmake default to the debug build. (bcc44f9f)
Updated the german translation. (3a95c2db)
Corrected the logic that locates the startup file and added some documentation. (51f2b9bf)