The commit id was always the same, fix this.
We previously used a way to recover the last tag from all branches.
However proceeding this way could leads in unattended behaviors.

First, imagine we are in the release 4.0, and a bug fix appears for 3.6
afterwardsi. The latest tag is not 4.0.x but 3.6.x and the package is not
updated.

Second, the manner we were taking the tag was taking its fixed commit with
it, not the latest commit available after the tag.

Now, we are just using a more common manner to recover the commit id: we
are prefixing it with the latest tag reachable from the **current** branch.