selinux-alpm-hook 0.1-3 update
Call restorecon in batches in order to improve performance. It it not
possible to pipe the list of package files into "| xargs restorecon -F"
because restorecon could fail sometimes (for example when there is an
invalid file on the filesystem which makes calls to getxattr() fail).

Implement a simple filter as a while loop which sanitizes the list of
packages files before running restorecon on it, in batches handled by
xargs.