Updated version (r4911.a1e94cc5 -> r4916.dba3d3b9).
Changelog
=========
Unbreak the windows build. (dba3d3b9)
Update the plot slider tooltip on manual moves of the plot slider. (817f2499)
This assert did cause many false positives => removing it. (fe858295)
An export bug that for some reason at all only appeared in the italian translation. (43a9c813)
If that was the sporadic-sigsev-on-copy problem on windows? (0ddff74d)