Removed NCCL support (it can be enabled by uncommenting the proper lines)
The newly released NCCL version 2 requires registration at the
NVIDIA developer webiste and a manual download. As this breaks
automated builds by AUR helpers, NCCL was disabled by default in
this package. It can still be enabled by uncommenting the proper
lines in the PKGBUILD file (instructions are at the top of the
PKGBUILD file).