Remove install script
As AUR users tinxy and tsuflux pointed out, the script is not necessary.
The package depends on xorg-font-utils, which in turn depends on
xorg-mkfontdir, which includes the xorg-mkfontdir.hook hook, which runs
the /usr/share/libalpm/scripts/xorg-mkfontdir script, which runs
mkfontscale and mkfontdir. Furthermore, the package depends directly on
fontconfig, which includes the fontconfig.hook hook, which runs
`fc-cache -s`. Together, these hooks completely cover the contents of
the install script, rendering it unnecessary.