Go back to cabal because stack can't install data files
Switching back to cabal due to a bug in stack (see jgm/gitit#599 and
commercialhaskell/stack/#848 ) that's impossible to cleanly work around.

A local .cabal/config file is used to avoid contamination from the
user's ~/.cabal/config file.  ghc-pristine prevents trouble caused by
the haskell-* packages, which lack the static libraries we need (that
was the original reason we switched to stack).

The presence of data files complicates this a lot: while "cabal copy"
does copy the data files of gitit, it doesn't copy the data files of
other dependent libraries (e.g. filestore).  Therefore, we need to copy
all the data files straight from the .cabal-sandbox.  To avoid conflicts
with other packages, we namespace all the data files under
/usr/share/gitit.  And to avoid copying the docs, we re-override docdir
to something else.  Cabal still installs the man pages and we can't
re-override that, so we just rm it during packaging.

Lastly, "cabal install" doesn't have a --destdir, and the --prefix has
to be the sandbox.  This means the binaries will definitely store the
wrong path for data files.  Thankfully, --enable-relocatable avoids
hardcoded paths entirely, so we just need to preserve relative paths.

At this point there's no reason to bother using "cabal copy", so we may
as well use "cabal install" and avoid the configure and build steps.
gitit only has one executable anyway!