version bump for 0.27.0 release
I don't really think this is necessary; mpv's releases are pretty
arbitrary and users of this package should just update whenever. But
bumping isn't any harder than complaining about semantics in the
comments when it's flagged so I'll just do it. Not sure if I'll continue
to do it in the future.