relax dependency
For MPD to speak MPRIS, just the general mpdris software is needed, not
a specifically new version from Git, so the dependency can be relaxed.
According to the Arch Linux VCS package guidelines [1] it can be
expected that the mpdris2-git package provides mpdris2, thus keeping all
package relations intact.

[1]: https://wiki.archlinux.org/index.php/VCS_package_guidelines