Move 'linuxns' to 'optdepends', minor tweaks
I've moved 'linuxns' to 'optdepends', in order to let the user choose
which kernel to use ('linux-zen' is an alternative, that lives in @extra,
so it is prebuilt, or the user might choose to built their own kernel
with USER_NS enabled).  It is not ideal, because it declares it as optional,
while it is not, but unfortunately 'depends' cannot handle alternate dependencies.
And the build time of the linuxns package is too big to ignore.

By adding a 'replaces' array, we allow for smooth transition for the users
who had installed the previous version, named simply 'concourse'.

Also, I've made liberal use of vars to reduce redundancy.