font update and source change
It bothers me on a deep moral level to see something named 2.0 and then
not get version bumps when it's updated. But that's what we're dealing
with here. I guess I'll just keep bumping the pkgrel since the official
version number is apparently not changing. There are more glyphs in this
one than the old version, so that's nice.