Version 1.2
[upstream]
This release introduces a syntax for direct imports `import (x, y) z`,
and makes scopes required for class methods. The focus on this release
was on cleaning up little internal issues and bugfixing. This release
is the first to use codecov to track code coverage.

The cleanups and testing done in this release are so that the next
release can be jammed with features.