Remove ffmpeg-full dependency
It seems that a ffmpeg build compiled with cuda support (like
ffmpeg-full) is not needed at all for giving cuda suppport in mpv.

It compiles fine with --enable-cuda-hwaccel when using ffmpeg from
the official repositories. And the resulting mpv binary can play
videos with cuda hwaccel decoding, needing just nvidia-utils
(libcuda.so) to be installed for this. Note that nvidia-340xx-utils
and nvidia-304xx-utils provides nvidia-utils (libcuda.so) for
legacy nvidia users. Not even the cuda package is needed for
playing videos with cuda hwaccel decoding.

This change can be reviewed if something breaks or does not
behave like the expected.