Perl/CPAN Module Convert::BaseN
This is one of the (optional) perl-dependencies of Koha: https://koha-community.org/
Start "./koha_perl_deps.pl -a" after downloading and unpacking
the koha-tarball for a list of all packages (required and optional).

This is with the cpan2aur-bugfix,
see https://rt.cpan.org/Public/Bug/Display.html?id=123153 for details.