allow building against upstream ffmpeg for now
Also apply a patch for the issue that caused this mess. See the comments
on the AUR page for some links to what's going on.

This is a temporary situation until things settle down. I'm not sure
what I'll do if mpv continues to reject upstream ffmpeg long-term, but I
probably won't be continuing to carry these patches.