update ffmpeg workaround
Can't just revert the commit anymore, as the surrounding code has
changed since. Add a patch with a simple workaround that will hopefully
not break too often.

PR #5033 has been merged(ish) upstream. Remove the patch.

mpv now requires git master ffmpeg; the latest release is not new
enough. This package probably should start depending on ffmpeg-git, but
I'm not switching it yet because that isn't provided by the current
ffmpeg-mpv-git package.

We'll just live with the _dirty in the version for now. If this goes on,
I'll strip it from the pkgver, but it will stay in the built binary as
long as it's true, to respect those handling bug reports upstream. Some
kind of option to avoid applying the patch when using ffmpeg-mpv should
be doable, though, so those who are following upstream's recommendations
can get a fully supported build without -dirty.